move test helpers under the view handler namespace
The test helpers are in fact only view handler helpers.  This simplifies
the number of files you need to remember to require and more accurately
organizes the test helpers.
