+++ lib/deas/view_handler.rb
@@ -115,6 +115,28 @@ module Deas
 
     end
 
+    module TestHelpers
+
+      def self.included(klass)
+        require 'rack/request'
+        require 'rack/response'
+        require 'deas/test_runner'
+      end
+
+      def test_runner(handler_class, args = nil)
+        args ||= {}
+        args[:request]  ||= Rack::Request.new({})
+        args[:response] ||= Rack::Response.new
+        args[:session]  ||= args[:request].session
+        TestRunner.new(handler_class, args)
+      end
+
+      def test_handler(handler_class, args = nil)
+        test_runner(handler_class, args).handler
+      end
+
+    end
+
   end
 
 end
