+++ test/unit/respond_with_proxy_tests.rb
@@ -2,7 +2,6 @@ require 'assert'
 require 'deas/respond_with_proxy'
 
 require 'deas/handler_proxy'
-require 'deas/test_helpers'
 require 'deas/url'
 require 'deas/view_handler'
 
@@ -25,7 +24,7 @@ class Deas::RespondWithProxy
   end
 
   class HandlerClassTests < UnitTests
-    include Deas::TestHelpers
+    include Deas::ViewHandler::TestHelpers
 
     desc "handler class"
     setup do
